==> R/00_doit.R <==
# Run all scripts ---------------------------------------------------------
source(file = "R/01_load.R")
source(file = "R/02_clean.R")
source(file = "R/03_augment.R")
source(file = "R/04_analysis_i.R")
source(file = "R/05_PCA.R")
source(file = "R/06_Kmeans.R")
source(file = "R/07_heatmap.R")
